+++ framework/model/src/main/java/is/codion/framework/model/EntityTableModel.java
@@ -27,12 +27,10 @@ import is.codion.common.state.State;
 import is.codion.common.value.Value;
 import is.codion.framework.db.EntityConnection;
 import is.codion.framework.db.EntityConnectionProvider;
-import is.codion.framework.domain.entity.ColorProvider;
 import is.codion.framework.domain.entity.Entities;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.EntityType;
-import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
 
 import java.util.Collection;
@@ -150,22 +148,6 @@ public interface EntityTableModel<E extends EntityEditModel> extends FilterModel
 	 * @return the {@link State} controlling whether this table model is editable
 	 */
 	State editable();
-
-	/**
-	 * @param row the row for which to retrieve the background color
-	 * @param attribute the attribute for which to retrieve the background color
-	 * @return an Object representing the background color for this row and attribute, specified by the row entity
-	 * @see EntityDefinition.Builder#backgroundColorProvider(ColorProvider)
-	 */
-	Object backgroundColor(int row, Attribute<?> attribute);
-
-	/**
-	 * @param row the row for which to retrieve the foreground color
-	 * @param attribute the attribute for which to retrieve the foreground color
-	 * @return an Object representing the foreground color for this row and attribute, specified by the row entity
-	 * @see EntityDefinition.Builder#foregroundColorProvider(ColorProvider)
-	 */
-	Object foregroundColor(int row, Attribute<?> attribute);
 
 	/**
 	 * Deletes the selected entities